RUTAS PARA EL ÁRBOL DE LINKS FUNCIONANDO
Se corrigio la ruta para la creación de productos creando el metodo
create en el controlador del producto.

+++ database/migrations/2017_03_06_131650_create_productos_table.php
@@ -15,7 +15,7 @@ class CreateProductosTable extends Migration
          Schema::create('productos', function (Blueprint $table) {
             $table->increments('id_producto');
             $table->string('nombre_producto');
-            $table->integer('precio_de_venta');
+            $table->bigInteger('precio_de_venta');
             $table->string('ingredientes_del_producto');
             $table->timestamps();
         });
